requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Dance==7.1.0
Authlib==1.3.1
python-dotenv==1.0.1
openai==1.47.0
email-validator==2.2.0
Werkzeug==3.0.4
Gunicorn==21.2.0
requests==2.32.3
psycopg[binary]==3.2.10
google-generativeai==0.7.2
bleach==6.2.0
